I moved the use cases, renamed some names and deleted some files
