--- Web Ui/src/application/views/assignmentsViews/useCases/createAssingmentsAdapter.ts
@@ -1,6 +1,0 @@
-import { AssignmentDataObject } from "../../../../modules/Assignments/domain/assignmentInterfaces";
-import { createAssignment } from "../../../../modules/Assignments/repository/assignment.API";
-
-export const createAssignmentsUseCase = async (assignmentData: AssignmentDataObject) => {
-    return await createAssignment(assignmentData);
-  }
