--- Web Ui/src/application/views/assignmentsViews/useCases/fetchAssignmentsApater.ts
@@ -1,5 +1,0 @@
-import { fetchAssignments } from "../../../../modules/Assignments/repository/assignment.API";
-
-export const fetchAssignmentsUseCase = async () => {
-  return await fetchAssignments();
-}
