--- Web Ui/src/application/views/view-1/assets/mytext.txt
@@ -1,1 +1,0 @@
-aqui van assets que se usan  solo en la vista actual
