+++ Web Ui/src/modules/Assignments/application/GetAssignments.ts
@@ -1,0 +1,5 @@
+import { fetchAssignments } from "../repository/assignment.API";
+
+export const fetchAssignmentsUseCase = async () => {
+  return await fetchAssignments();
+}
