+++ Web Ui/src/modules/Assignments/application/createAssingmentsAdapter.ts
@@ -1,0 +1,6 @@
+import { AssignmentDataObject } from "../domain/assignmentInterfaces";
+import { createAssignment } from "../repository/assignment.API";
+
+export const createAssignmentsUseCase = async (assignmentData: AssignmentDataObject) => {
+    return await createAssignment(assignmentData);
+  }
