+++ Web Ui/src/sections/Assignments/components/AssignmentForm.tsx
@@ -6,7 +6,7 @@ import { Box, Container, TextField, Typography } from "@mui/material";
 import { LocalizationProvider } from "@mui/x-date-pickers";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import Filter from "./DatePicker";
-import { createAssignmentsUseCase } from "../../../application/views/assignmentsViews/useCases/createAssingmentsAdapter";
+import { createAssignmentsUseCase } from "../../../modules/Assignments/application/createAssingmentsAdapter";
 
 function Formulario() {
   const [successMessage, setSuccessMessage] = useState<string | null>(null);
